requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
cffi==1.14.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
jwt==1.2.0
kombu==5.1.0
Markdown==3.3.4
numpy==1.21.1
Pillow==8.3.1
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
